+++ packages/start-static-server-functions/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @tanstack/start-static-server-functions
 
+## 1.167.34
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/start-client-core@1.170.29
+  - @tanstack/react-start@1.168.51
+  - @tanstack/solid-start@1.168.49
+
 ## 1.167.33
 
 ### Patch Changes